Moves xdev and xint functions to rcpp_xint_xdev_functions. Removes get_bin_taxonomy_report #36
