+++ src/RcppExports.cpp
@@ -287,18 +287,6 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// get_bin_taxonomy_report
-Rcpp::DataFrame get_bin_taxonomy_report(Rcpp::Environment data, string type);
-RcppExport SEXP _rdataset_get_bin_taxonomy_report(SEXP dataSEXP, SEXP typeSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< Rcpp::Environment >::type data(dataSEXP);
-    Rcpp::traits::input_parameter< string >::type type(typeSEXP);
-    rcpp_result_gen = Rcpp::wrap(get_bin_taxonomy_report(data, type));
-    return rcpp_result_gen;
-END_RCPP
-}
 // get_bin_types
 vector<string> get_bin_types(Rcpp::Environment data);
 RcppExport SEXP _rdataset_get_bin_types(SEXP dataSEXP) {
@@ -885,7 +873,6 @@ static const R_CallMethodDef CallEntries[] = {
     {"_rdataset_get_bin_abundances", (DL_FUNC) &_rdataset_get_bin_abundances, 3},
     {"_rdataset_get_bin_names", (DL_FUNC) &_rdataset_get_bin_names, 2},
     {"_rdataset_get_bin_representative_sequences", (DL_FUNC) &_rdataset_get_bin_representative_sequences, 2},
-    {"_rdataset_get_bin_taxonomy_report", (DL_FUNC) &_rdataset_get_bin_taxonomy_report, 2},
     {"_rdataset_get_bin_types", (DL_FUNC) &_rdataset_get_bin_types, 1},
     {"_rdataset_get_dataset_name", (DL_FUNC) &_rdataset_get_dataset_name, 1},
     {"_rdataset_get_list", (DL_FUNC) &_rdataset_get_list, 2},
