+++ src/rcpp_dataset.cpp
@@ -1,54 +1,8 @@
 #include <Rcpp.h>
 #include "../inst/include/rdataset.h"
 #include "dataset.h"
+#include "rcpp_xint_xdev_functions.h"
 
-/******************************************************************************/
-SEXP fill_required_parameters(const Rcpp::DataFrame df,
-                              const string& given_column_name,
-                              string type = "string") {
-
-    if (!df.containsElementNamed(given_column_name.c_str())) {
-        string message = "Expected a data.frame column named " +
-            given_column_name +" to be provided.";
-
-        throw Rcpp::exception(message.c_str());
-    }
-
-    return df[given_column_name];
-}
-/******************************************************************************/
-SEXP fill_optional_parameters(const Rcpp::DataFrame df,
-                              const string& default_column_name,
-                              const string& given_column_name,
-                              string type = "string") {
-
-
-        // Check if the column exists in the data frame
-        if (df.containsElementNamed(given_column_name.c_str())) {
-            return df[given_column_name];
-        }else if (df.containsElementNamed(default_column_name.c_str())) {
-            return df[default_column_name];
-        }else if ((given_column_name == "") ||
-                        (given_column_name == default_column_name)) {
-            //fall through
-        }else {
-            string message = "Expected a data.frame column named " +
-                given_column_name +" to be provided.";
-
-            throw Rcpp::exception(message.c_str());
-        }
-
-    if (type == "string") {
-        return Rcpp::CharacterVector(0);
-    }else if ((type == "float") || (type == "double")) {
-        return Rcpp::NumericVector(0);
-    }else{
-        string message = "Unsupported column type for conversion.";
-        throw Rcpp::exception(message.c_str());
-    }
-
-    return Rcpp::CharacterVector(0);
-}
 /******************************************************************************/
 //' @title new_dataset
 //' @description
@@ -219,7 +173,7 @@ void add_report(Rcpp::Environment data,
 
             // sequence names in table
             vector<string> sequenceNames = Rcpp::as<vector<string>>(
-                fill_required_parameters(table, sequence_name));
+                xint_fill_required_parameters(table, sequence_name));
 
             // if we don't have any sequences yet, add them
             if (d.get()->getTotal("") == 0) {
@@ -307,18 +261,18 @@ double add_references(Rcpp::Environment data,
     vector<string> reference_names, reference_versions, reference_usages;
     vector<string> reference_notes, reference_urls;
 
-    reference_names = Rcpp::as<vector<string>>(fill_required_parameters(table,
+    reference_names = Rcpp::as<vector<string>>(xint_fill_required_parameters(table,
                                                                         reference_name));
-    reference_versions = Rcpp::as<vector<string>>(fill_optional_parameters(table,
+    reference_versions = Rcpp::as<vector<string>>(xint_fill_optional_parameters(table,
                                                                   "reference_versions",
                                                                   reference_version));
-    reference_usages = Rcpp::as<vector<string>>(fill_optional_parameters(table,
+    reference_usages = Rcpp::as<vector<string>>(xint_fill_optional_parameters(table,
                                                                            "reference_usages",
                                                                            reference_usage));
-    reference_notes = Rcpp::as<vector<string>>(fill_optional_parameters(table,
+    reference_notes = Rcpp::as<vector<string>>(xint_fill_optional_parameters(table,
                                                                            "reference_notes",
                                                                            reference_note));
-    reference_urls = Rcpp::as<vector<string>>(fill_optional_parameters(table,
+    reference_urls = Rcpp::as<vector<string>>(xint_fill_optional_parameters(table,
                                                                        "reference_urls",
                                                                        reference_url));
 
@@ -416,11 +370,11 @@ double add_sequences(Rcpp::Environment data,
 
     vector<string> sequence_names, sequences, comments;
     sequence_names = Rcpp::as<vector<string>>(
-        fill_required_parameters(table, sequence_name));
-    sequences = Rcpp::as<vector<string>>(fill_optional_parameters(table,
+        xint_fill_required_parameters(table, sequence_name));
+    sequences = Rcpp::as<vector<string>>(xint_fill_optional_parameters(table,
                                                                   "sequences",
                                                                   sequence));
-    comments = Rcpp::as<vector<string>>(fill_optional_parameters(table,
+    comments = Rcpp::as<vector<string>>(xint_fill_optional_parameters(table,
                                                                   "comments",
                                                                   comment));
 
@@ -510,15 +464,15 @@ double assign_bins(Rcpp::Environment data,
     vector<float> abundances;
 
     // fill vectors with columns from table
-    bin_names = Rcpp::as<vector<string>>(fill_required_parameters(table,
+    bin_names = Rcpp::as<vector<string>>(xint_fill_required_parameters(table,
                                                                   bin_name));
-    samples = Rcpp::as<vector<string>>(fill_optional_parameters(table,
+    samples = Rcpp::as<vector<string>>(xint_fill_optional_parameters(table,
                                                                 "samples",
                                                                 sample));
-    sequence_names = Rcpp::as<vector<string>>(fill_optional_parameters(table,
+    sequence_names = Rcpp::as<vector<string>>(xint_fill_optional_parameters(table,
                                                                  "sequence_names",
                                                                  sequence_name));
-    abundances = Rcpp::as<vector<float>>(fill_optional_parameters(table,
+    abundances = Rcpp::as<vector<float>>(xint_fill_optional_parameters(table,
                                                                   "abundances",
                                                                   abundance,
                                                                   "float"));
@@ -625,9 +579,9 @@ double assign_bin_representative_sequences(Rcpp::Environment data,
                                            string sequence_name = "sequence_names") {
 
     vector<string> bin_names,  sequence_names;
-    bin_names = Rcpp::as<vector<string>>(fill_required_parameters(table,
+    bin_names = Rcpp::as<vector<string>>(xint_fill_required_parameters(table,
                                                                   bin_name));
-    sequence_names = Rcpp::as<vector<string>>(fill_required_parameters(table,
+    sequence_names = Rcpp::as<vector<string>>(xint_fill_required_parameters(table,
                                                                        sequence_name));
 
     Rcpp::XPtr<Dataset> d = data["data"];
@@ -707,9 +661,9 @@ double assign_bin_taxonomy(Rcpp::Environment data,
     }
 
     vector<string> bin_names, taxonomies;
-    bin_names = Rcpp::as<vector<string>>(fill_required_parameters(table,
+    bin_names = Rcpp::as<vector<string>>(xint_fill_required_parameters(table,
                                                                   bin_name));
-    taxonomies = Rcpp::as<vector<string>>(fill_required_parameters(table,
+    taxonomies = Rcpp::as<vector<string>>(xint_fill_required_parameters(table,
                                                                    taxonomy));
 
     double numAssigned = d.get()->assignBinTaxonomy(bin_names,
@@ -779,15 +733,15 @@ double assign_sequence_abundance(Rcpp::Environment data,
     vector<string> sequence_names, samples, treatments;
     vector<float> abundances;
 
-    sequence_names = Rcpp::as<vector<string>>(fill_required_parameters(table,
+    sequence_names = Rcpp::as<vector<string>>(xint_fill_required_parameters(table,
                                                                        sequence_name));
-    samples = Rcpp::as<vector<string>>(fill_optional_parameters(table,
+    samples = Rcpp::as<vector<string>>(xint_fill_optional_parameters(table,
                                                                 "samples",
                                                                 sample));
-    treatments = Rcpp::as<vector<string>>(fill_optional_parameters(table,
+    treatments = Rcpp::as<vector<string>>(xint_fill_optional_parameters(table,
                                                                 "treatments",
                                                                 treatment));
-    abundances = Rcpp::as<vector<float>>(fill_optional_parameters(table,
+    abundances = Rcpp::as<vector<float>>(xint_fill_optional_parameters(table,
                                                                    "abundances",
                                                                    abundance,
                                                                    "float"));
@@ -874,9 +828,9 @@ double assign_sequence_taxonomy(Rcpp::Environment data,
                                 string taxonomy = "taxonomies") {
 
     vector<string> sequence_names, taxonomies;
-    sequence_names = Rcpp::as<vector<string>>(fill_required_parameters(table,
+    sequence_names = Rcpp::as<vector<string>>(xint_fill_required_parameters(table,
                                                                        sequence_name));
-    taxonomies = Rcpp::as<vector<string>>(fill_required_parameters(table,
+    taxonomies = Rcpp::as<vector<string>>(xint_fill_required_parameters(table,
                                                                     taxonomy));
     // make sure names is same size as taxonomies
     if (sequence_names.size() != taxonomies.size()) {
@@ -948,9 +902,9 @@ double assign_treatments(Rcpp::Environment data,
                          string treatment = "treatments") {
 
     vector<string> samples, treatments;
-    samples = Rcpp::as<vector<string>>(fill_required_parameters(table,
+    samples = Rcpp::as<vector<string>>(xint_fill_required_parameters(table,
                                                                 sample));
-    treatments = Rcpp::as<vector<string>>(fill_required_parameters(table,
+    treatments = Rcpp::as<vector<string>>(xint_fill_required_parameters(table,
                                                                    treatment));
 
     // check to make sure samples and treatments are same length
@@ -1213,25 +1167,6 @@ Rcpp::DataFrame get_bin_representative_sequences(Rcpp::Environment data,
                                                  string type = "otu") {
     Rcpp::XPtr<Dataset> d = data["data"];
     return d.get()->getBinRepresentativeSequences(type);
-}
-/******************************************************************************/
-//' @title get_bin_taxonomy_report
-//' @description
-//' Get the bin classifications of a \link{dataset} object
-//'
-//' @param data, a \link{dataset} object
-//' @param type, string indicating the type of bin clusters. Default = "otu".
-//' @examples
-//'
-//' data <- miseq_sop_example()
-//' get_bin_taxonomy_report(data, "otu")
-//'
-//' @return data.frame
-//[[Rcpp::export]]
-Rcpp::DataFrame get_bin_taxonomy_report(Rcpp::Environment data,
-                                        string type = "otu") {
-    Rcpp::XPtr<Dataset> d = data["data"];
-    return d.get()->getBinTaxonomyReport(type);
 }
 /******************************************************************************/
 //' @title get_bin_types
@@ -1960,590 +1895,4 @@ bool is_aligned(Rcpp::Environment data) {
      return d.get()->isAligned;
 }
 /******************************************************************************/
-//' @title xdev_merge_bins
-//' @description
-//' Designed with package integration in mind, the merge bins function allows
-//' you to merge bins in a \link{dataset} object
-//'
-//' @param data, a \link{dataset} object.
-//'
-//' @param bin_names, a vector of strings containing the names of the bins you
-//' would like merge. The resulting merged bin will be stored in the first
-//' bin_id in the vector.
-//' @param reason, a string indicating why you are merging bins. Default =
-//' "merged".
-//' @param type, a string indicating the type of bin clusters. Default = "otu"
-//'
-//' @examples
-//'
-//'  data <- miseq_sop_example()
-//'
-//'  # to merge otu5 and otu6
-//'
-//'  bins_to_merge <- c("Otu005", "Otu006")
-//'
-//'  xdev_merge_bins(data, bins_to_merge)
-//'
-//'  # If you look at the scrap report, you will see Otu006 with the trash code
-//'  # set to "merged".
-//'
-//'  report(data, "bin_scrap")
-//'
-//[[Rcpp::export]]
-void xdev_merge_bins(Rcpp::Environment data, vector<string> bin_names,
-                string reason = "merged", string type = "otu") {
-    Rcpp::XPtr<Dataset> d = data["data"];
-    d.get()->mergeBins(bin_names, reason, type);
-}
-/******************************************************************************/
-//' @title xdev_merge_sequences
-//' @description
-//' Designed with package integration in mind, the merge sequences function
-//' allows you to merge sequences in a \link{dataset} object.
-//'
-//' @param data, a \link{dataset} object.
-//'
-//' @param sequence_names, a vector of strings containing the names of the
-//' sequences you would like merge. The resulting merged sequence will be stored
-//' in the first sequence name in the vector.
-//' @param reason a string indicating why you are merging sequences.
-//' Default = "merged"
-//'
-//' @examples
-//'
-//' data <- miseq_sop_example()
-//'
-//' get_num_sequences(data)
-//'
-//' # For the sake of example let's merge the first 3 sequences from
-//' # miseq_sop_example:
-//'
-//' seqs_to_merge <- c("M00967_43_000000000-A3JHG_1_2101_16474_12783",
-//'                    "M00967_43_000000000-A3JHG_1_1113_12711_3318",
-//'                    "M00967_43_000000000-A3JHG_1_2108_14707_9807")
-//'
-//' xdev_merge_sequences(data, seqs_to_merge)
-//'
-//' # If you look at the scrap report, you will see the second two sequence
-//' # names, listed with the trash code set to "merged".
-//'
-//' report(data, "sequence_scrap")
-//'
-//' # You can see from the get_num_sequences function that the merged sequence's
-//' # abundances are added to the first sequence.
-//'
-//' get_num_sequences(data)
-//'
-//[[Rcpp::export]]
-void xdev_merge_sequences(Rcpp::Environment data, vector<string> sequence_names,
-                 string reason = "merged") {
-    Rcpp::XPtr<Dataset> d = data["data"];
-    d.get()->mergeSequences(sequence_names, reason);
-}
-/******************************************************************************/
-//' @title xdev_remove_bins
-//' @description
-//' Designed with package integration in mind, the remove bins function allows
-//' you to remove bins from a \link{dataset} object
-//'
-//' @param data, a \link{dataset} object.
-//'
-//' @param bin_names, a vector of strings containing the names of the bins you
-//' would like removed.
-//' @param trash_tags, a vector of strings containing the reasons you are
-//' removing each bin
-//' @param type a string indicating the type of clusters.
-//' @examples
-//'
-//'   data <- new_dataset("my_dataset", 2)
-//'
-//'   bin_names <- c("bin1", "bin2", "bin3")
-//'   abundances <- c(110, 525, 80)
-//'
-//'   assign_bins(data, data.frame(bin_names = bin_names,
-//'                                abundances = abundances))
-//'
-//'   get_num_bins(data)
-//'
-//'   bins_to_remove <- c("bin1")
-//'   trash_tag <- c("bad_bin")
-//'
-//'   xdev_remove_bins(data, bins_to_remove, trash_tag)
-//'
-//'   get_num_bins(data)
-//'
-//[[Rcpp::export]]
-void xdev_remove_bins(Rcpp::Environment data, vector<string> bin_names,
-                 vector<string> trash_tags, string type = "otu") {
-    Rcpp::XPtr<Dataset> d = data["data"];
-    d.get()->removeBins(bin_names, trash_tags, type);
-}
-/******************************************************************************/
-//' @title xdev_remove_lineages
-//' @description
-//' Designed with package integration in mind, the remove lineages function
-//' allows you to remove contaminents from a \link{dataset}
-//'
-//' @param data, a \link{dataset} object.
-//'
-//' @param contaminants, vector of strings containing the taxonomies you would
-//' like to remove
-//' @param trash_tag, a string containing reason you are removing the lineages.
-//' Default = "contaminant".
-//'
-//' @examples
-//' data <- read_mothur(fasta = rdataset_example("final.fasta"),
-//'                        count = rdataset_example("final.count_table"),
-//'                        taxonomy = rdataset_example("final.taxonomy"),
-//'                        design = rdataset_example("mouse.time.design"),
-//'                        otu_list = rdataset_example("final.opti_mcc.list"),
-//'                        dataset_name = "miseq_sop")
-//'
-//' contaminants <- c("Chloroplast", "Mitochondria", "unknown", "Archaea",
-//'  "Eukaryota")
-//'
-//' xdev_remove_lineages(data, contaminants)
-//'
-//[[Rcpp::export]]
-void xdev_remove_lineages(Rcpp::Environment data, vector<string> contaminants,
-                     string trash_tag = "contaminant") {
-    Rcpp::XPtr<Dataset> d = data["data"];
-    d.get()->removeLineages(contaminants, trash_tag);
-}
-/******************************************************************************/
-//' @title xdev_remove_samples
-//' @description
-//' Designed with package integration in mind, the remove samples function allows
-//' you to remove samples from a \link{dataset} object
-//'
-//' @param data, a \link{dataset} object.
-//'
-//' @param samples, vector of strings containing the names of the samples to
-//' remove.
-//'
-//' @examples
-//'
-//' data <- miseq_sop_example()
-//'
-//' get_num_samples(data)
-//'
-//' # To remove samples 'F3D0' and 'F3D1'
-//'
-//' xdev_remove_samples(data, c("F3D0", "F3D1"))
-//'
-//' get_num_samples(data)
-//'
-//[[Rcpp::export]]
-void xdev_remove_samples(Rcpp::Environment data, vector<string> samples) {
-    Rcpp::XPtr<Dataset> d = data["data"];
-    d.get()->removeSamples(samples);
-}
-/******************************************************************************/
-//' @title xdev_remove_sequences
-//' @description
-//' Designed with package integration in mind, the remove sequences function
-//' allows you to remove sequences from a \link{dataset} object
-//'
-//' @param data, a \link{dataset} object.
-//'
-//' @param sequence_names, vector of strings containing the names of the
-//' sequences to remove
-//' @param trash_tags vector of strings containing the reasons for the sequences
-//' removals
-//'
-//' @examples
-//'
-//' data <- miseq_sop_example()
-//'
-//' get_num_sequences(data)
-//'
-//' # For the sake of example let's remove the first 3 sequences from
-//' # miseq_sop_example:
-//'
-//' seqs_to_remove <- c("M00967_43_000000000-A3JHG_1_2101_16474_12783",
-//'                    "M00967_43_000000000-A3JHG_1_1113_12711_3318",
-//'                    "M00967_43_000000000-A3JHG_1_2108_14707_9807")
-//' trash_codes <- c("example", "removing", "sequences")
-//'
-//' xdev_remove_sequences(data, seqs_to_remove, trash_codes)
-//'
-//' # If you look at the scrap report, you the sequences names, listed with the
-//' # trash codes set to "example", "removing", "sequences".
-//'
-//' report(data, "sequence_scrap")
-//'
-//' # You can see from the get_num_sequences function that the removed
-//' # sequence's abundances are removed from the dataset.
-//'
-//' get_num_sequences(data)
-//'
-//[[Rcpp::export]]
-void xdev_remove_sequences(Rcpp::Environment data,
-                      vector<string> sequence_names,
-                      vector<string> trash_tags) {
 
-    Rcpp::XPtr<Dataset> d = data["data"];
-    d.get()->removeSequences(sequence_names, trash_tags);
-}
-/******************************************************************************/
-//' @title xdev_set_abundance
-//' @description
-//' Designed with package integration in mind, the set abundance function
-//' allows you to change the abundances of sequences in a \link{dataset} object
-//' without samples.
-//'
-//' @param data, a \link{dataset} object
-//'
-//' @param sequence_names, a vector of strings containing sequence names
-//' @param sequence_abundances, vector containing the abundances of each
-//' sequence.
-//' @param reason, a string containing the trash tag to be applied to any
-//' sequences set to 0 abundance. Default = "update".
-//'
-//' @examples
-//'
-//' names <- c("seq1", "seq2", "seq3",  "seq4")
-//' abunds <- c(1250, 65, 50, 4)
-//'
-//' data <- new_dataset("my_dataset", 2)
-//'
-//' assign_sequence_abundance(data, data.frame(sequence_names = names,
-//'                                            abundances = abunds))
-//' get_sequence_abundances(data)
-//'
-//' seqs_to_update <- c("seq1", "seq3")
-//' new_abunds <- c(1000, 100)
-//'
-//' xdev_set_abundance(data, seqs_to_update, new_abunds)
-//'
-//' get_sequence_abundances(data)
-//'
-//[[Rcpp::export]]
-void xdev_set_abundance(Rcpp::Environment data,
-                   vector<string> sequence_names,
-                   vector<float> sequence_abundances,
-                   string reason = "update") {
-
-    Rcpp::XPtr<Dataset> d = data["data"];
-
-    if (get_num_samples(data) == 0) {
-        d.get()->setAbundance(sequence_names, sequence_abundances, reason);
-    }else{
-        string message = "[ERROR]: You cannot set the total sequence abundance";
-        message += " for sequences whose abundances are parsed by sample. ";
-        message += "Try 'set_abundances' instead of 'set_abundance'.";
-        RcppThread::Rcerr << endl << message << endl;
-        throw Rcpp::exception(message.c_str());
-    }
-}
-/******************************************************************************/
-//' @title xdev_set_abundances
-//' @description
-//' Designed with package integration in mind, the set abundances function
-//' allows you to change the abundances of sequences in a \link{dataset} object
-//' with samples.
-//'
-//' @param data, a \link{dataset} object
-//'
-//' @param sequence_names, a vector of strings containing sequence names
-//' @param abundances, 2D vector ([num_seqs][num_samples]) containing
-//' the abundances of each sequence parsed by sample.
-//' @param reason, a string containing the trash tag to be applied to any
-//' sequences set to 0 abundance. Default = "update".
-//' @examples
-//'
-//' data <- new_dataset("my_dataset", 2)
-//'
-//' sequence_names <- c("seq1", "seq1", "seq1", "seq2", "seq2", "seq2", "seq3",
-//'                     "seq3", "seq4")
-//' samples <- c("sample2", "sample3", "sample4", "sample2", "sample3",
-//'              "sample4", "sample2", "sample3", "sample4")
-//' abundances <- c(250, 400, 500, 25, 40, 50, 25, 25, 4)
-//'
-//' assign_sequence_abundance(data, data.frame(sequence_names = sequence_names,
-//'                                            abundances = abundances,
-//'                                            samples = samples))
-//'
-//' seqs_to_update <- c("seq4")
-//' new_abunds <- list(c(20, 10, 4))
-//'
-//' xdev_set_abundances(data, seqs_to_update, new_abunds)
-//'
-//[[Rcpp::export]]
-void xdev_set_abundances(Rcpp::Environment data,
-                   vector<string> sequence_names,
-                   vector<vector<float>> abundances,
-                   string reason = "update") {
-
-    Rcpp::XPtr<Dataset> d = data["data"];
-
-    if (get_num_samples(data) != 0) {
-        d.get()->setAbundances(sequence_names, abundances, reason);
-    }else {
-        string message = "[ERROR]: You cannot set parsed sequence abundances ";
-        message += "when your dataset does not include sample data. ";
-        message += "Try 'set_abundance' instead of 'set_abundances'.";
-        RcppThread::Rcerr << endl << message << endl;
-        throw Rcpp::exception(message.c_str());
-    }
-}
-/******************************************************************************/
-//' @title xdev_set_bin_abundance
-//' @description
-//' Designed with package integration in mind, the set bin abundance function
-//' allows you to change the abundances of bins in a \link{dataset} object
-//' without sample data.
-//'
-//' @param data, a \link{dataset} object
-//'
-//' @param bin_names, a vector strings containing of bin names to set the
-//' abundances for.
-//' @param abundances, vector containing the abundances of each bin.
-//' @param type, a string indicating the type of clusters. Default = "otu".
-//' @param reason, a string containing the trash tag to be applied to any bins
-//'  set to 0 abundance. Default = "update".
-//'
-//' @examples
-//'   # For example sake, let's create a dataset with 3 bins:
-//'
-//'   data <- new_dataset("my_dataset", 2)
-//'
-//'   bin_ids <- c("bin1", "bin2", "bin3")
-//'   abundances <- c(110, 525, 80)
-//'
-//'   assign_bins(data, data.frame(bin_names = bin_ids,
-//'                                abundances = abundances))
-//'
-//'   get_bin_abundance(data, "bin1")
-//'   get_bin_abundance(data, "bin2")
-//'
-//'   # Now we can use set_bin_abundance to change the abundances of bin1 and
-//'   # bin2
-//'
-//'   bins <- c("bin1", "bin2")
-//'   new_abunds <- c(300, 250)
-//'
-//'   xdev_set_bin_abundance(data, bins, new_abunds)
-//'
-//'   get_bin_abundance(data, "bin1")
-//'   get_bin_abundance(data, "bin2")
-//'
-//[[Rcpp::export]]
-void xdev_set_bin_abundance(Rcpp::Environment data,
-                       vector<string> bin_names,
-                       vector<float> abundances,
-                       string type = "otu",
-                       string reason = "update") {
-
-    Rcpp::XPtr<Dataset> d = data["data"];
-
-    if (d.get()->hasListAssignments(type)) {
-        string message = "[ERROR]: You cannot set the bin abundance for bin ";
-        message += "clusters with sequence assignments.";
-        RcppThread::Rcerr << endl << message << endl;
-        throw Rcpp::exception(message.c_str());
-
-    }else if (get_num_samples(data) != 0) {
-        string message = "[ERROR]: You cannot set the total bin abundance";
-        message += " for bins whose abundances are parsed by sample. ";
-        message += "Try 'set_bin_abundances' instead of 'set_bin_abundance'.";
-        RcppThread::Rcerr << endl << message << endl;
-        throw Rcpp::exception(message.c_str());
-    }else{
-        d.get()->setBinAbundance(bin_names, abundances, reason, type);
-    }
-}
-/******************************************************************************/
-//' @title xdev_set_bin_abundances
-//' @description
-//' Designed with package integration in mind, the set bin abundances function
-//' allows you to change the abundances of bins in a \link{dataset} object
-//' with sample data.
-//'
-//' @param data, a \link{dataset} object
-//'
-//' @param bin_names, a vector strings containing of bin names to set the
-//' abundances for.
-//' @param abundances, 2D vector ([num_seqs][num_samples]) containing the
-//' abundances of each bin parsed by sample.
-//' @param type a string indicating the type of clusters. Default = "otu".
-//' @param reason, a string containing the trash tag to be applied to any bins
-//'  set to 0 abundance. Default = "update".
-//'
-//' @examples
-//'
-//'   # For example sake, let's create a dataset with 3 bins:
-//'
-//'   data <- new_dataset("my_dataset", 2)
-//'
-//'   bin_ids <- c("bin1", "bin1", "bin1", "bin2", "bin2", "bin3")
-//'   samples <- c("sample1", "sample2", "sample5", "sample1", "sample3",
-//'                "sample1")
-//'   sample_abundances <- c(10, 100, 1, 500, 25, 80)
-//'
-//'   assign_bins(data, data.frame(bin_names = bin_ids,
-//'                                abundances = sample_abundances,
-//'                                samples = samples))
-//'
-//'   # You can see bin1's abundances parsed by sample using get_bin_abundances:
-//'   get_bin_abundances(data, "bin1")
-//'
-//'   # You can change bin1's abundances as follows:
-//'
-//'   new_bin1_abunds <- list(c(10,50,0,0))
-//'   bins <- c("bin1")
-//'
-//'   xdev_set_bin_abundances(data, bins, new_bin1_abunds)
-//'
-//'   get_bin_abundances(data, "bin1")
-//'
-//[[Rcpp::export]]
-void xdev_set_bin_abundances(Rcpp::Environment data,
-                        vector<string> bin_names,
-                        vector<vector<float>> abundances,
-                        string type = "otu", string reason = "update") {
-
-    Rcpp::XPtr<Dataset> d = data["data"];
-
-    if (d.get()->hasListAssignments(type)) {
-        string message = "[ERROR]: You cannot set the bin abundance for bin ";
-        message += "clusters with sequence assignments.";
-        RcppThread::Rcerr << endl << message << endl;
-        throw Rcpp::exception(message.c_str());
-    }else if (get_num_samples(data) == 0) {
-        string message = "[ERROR]: You cannot set parsed bin abundances ";
-        message += "when your dataset does not include sample data. ";
-        message += "Try 'set_bin_abundance' instead of 'set_bin_abundances'.";
-        RcppThread::Rcerr << endl << message << endl;
-        throw Rcpp::exception(message.c_str());
-    }else{
-        d.get()->setBinAbundances(bin_names, abundances, reason, type);
-    }
-}
-/******************************************************************************/
-//' @title xdev_set_sequences
-//' @description
-//' Designed with package integration in mind, the set sequences function
-//' allows you to change the nucleotide strings of sequences in a \link{dataset}
-//' object. For example, set_sequences may be used after alignment to overwrite
-//' the unaligned sequences with aligned sequences.
-//'
-//' @param data, a \link{dataset} object
-//' @param sequence_names, a vector of strings containing sequence names
-//' @param sequences, a vector of strings containing sequence nucleotide strings
-//' @param comments, a vector of strings containing sequence comments.
-//' (Optional)
-//'
-//' @examples
-//'
-//' data <- new_dataset("my_dataset", 2)
-//'
-//' add_sequences(data, data.frame(sequence_names = c("seq1", "seq2",
-//'                                                   "seq3", "seq4")))
-//'
-//' xdev_set_sequences(data, c("seq1", "seq2","seq3", "seq4"),
-//'                     c("ATTGC", "ACTGC", "AGTGC", "TTTGC"))
-//'
-//[[Rcpp::export]]
-void xdev_set_sequences(Rcpp::Environment data,
-                   vector<string> sequence_names,
-                   vector<string> sequences,
-                   Rcpp::CharacterVector comments = Rcpp::CharacterVector::create()) {
-
-    Rcpp::XPtr<Dataset> d = data["data"];
-    d.get()->setSequences(sequence_names, sequences,
-                          Rcpp::as<vector<string>>(comments));
-}
-/******************************************************************************/
-//' @title xdev_set_dataset_name
-//' @description
-//' Designed with package integration in mind, set the name of a \link{dataset} object.
-//'
-//' @param data, a \link{dataset} object
-//' @param dataset_name, a string containing the desired name
-//'
-//' @examples
-//'
-//' data <- new_dataset("my_dataset", 2)
-//' xdev_set_dataset_name(data, "new_dataset_name")
-//'
-//[[Rcpp::export]]
-void xdev_set_dataset_name(Rcpp::Environment data, string dataset_name) {
-    Rcpp::XPtr<Dataset> d = data["data"];
-    d.get()->datasetName = dataset_name;
-}
-/******************************************************************************/
-//' @title xdev_set_num_processors
-//' @description
-//' Designed with package integration in mind, set the number of processors used
-//'  to summarize a \link{dataset} object
-//'
-//' @param data, a \link{dataset} object
-//' @param processors, a integer containing the desired number of processors
-//' @examples
-//'
-//' data <- new_dataset("my_dataset", 2)
-//' xdev_set_num_processors(data, 1)
-//'
-//[[Rcpp::export]]
-void xdev_set_num_processors(Rcpp::Environment data, int processors) {
-    Rcpp::XPtr<Dataset> d = data["data"];
-    d.get()->processors = processors;
-}
-/******************************************************************************/
-
-/******************************************************************************/
-//' @title xint_copy_pointer
-//' @name xint_copy_pointer
-//' @description
-//' For internal use only, copy an instance of the C++ 'Dataset' class.
-//' @param data, a \link{dataset} object
-//' @return pointer to an instance of the C++ 'Dataset' class.
-//' @keywords internal
-//[[Rcpp::export]]
-Rcpp::XPtr<Dataset> xint_copy_pointer(Rcpp::Environment data) {
-    Rcpp::XPtr<Dataset> d = data["data"];
-    Dataset* copy = new Dataset(*d.get());
-    return Rcpp::XPtr<Dataset>(copy);
-}
-/******************************************************************************/
-//' @title xint_new_pointer
-//' @name xint_new_pointer
-//' @description
-//' For internal use only, create an instance of the C++ 'Dataset' class.
-//' @param dataset_name, string containing dataset name
-//' @param processors, number of processors to use
-//' @return pointer to an instance of the C++ 'Dataset' class.
-//' @keywords internal
-//[[Rcpp::export]]
-Rcpp::XPtr<Dataset> xint_new_pointer(string dataset_name = "", int processors = 1) {
-     Dataset* d = new Dataset(dataset_name, processors);
-     return Rcpp::XPtr<Dataset>(d);
-}
-/******************************************************************************/
-//' @title xint_deserialize_dobject
-//' @name xint_deserialize_dobject
-//' @description
-//' For internal use only, deserialize_dobject an instance of the C++ 'Dataset'
-//'  class.
-//' @param data, a \link{dataset} object
-//[[Rcpp::export]]
-void xint_deserialize_dobject(Rcpp::Environment data) {
-     data["data"] = xint_new_pointer();
-     Rcpp::XPtr<Dataset> d = data["data"];
-     const Rcpp::RawVector raw = data["raw"];
-     d.get()->loadFromSerialized(raw);
-}
-/******************************************************************************/
-//' @title xint_serialize_dobject
-//' @name xint_serialize_dobject
-//' @description
-//' For internal use only, xint_serialize_dobject an instance of the C++ 'Dataset'
-//' class.
-//' @param data, a \link{dataset} object
-//[[Rcpp::export]]
-void xint_serialize_dobject(Rcpp::Environment data) {
-     Rcpp::XPtr<Dataset> d = data["data"];
-     data["raw"] = d.get()->serializeDataset();
- }
-/******************************************************************************/
